init: Force a log to the kernel log for boot animation complete
Need to force a log statement to the kernel log to determine when the
boot animation is completed.  Boot animation exit is when the UI is ready
for display to the user.

If the logcat output is redirected to kmsg there is a 6 second addition to
overall boot time.  This is obviously not acceptable.

There is always boot chart but that requires a rebuild and a special know how on
how to do that so that should be relugated to developers.

So that leaves us with being able to replicate this from just a kernel log without
adding the overhead of special tools or code.

Change-Id: I3d1c1692422f1701cb0c55e34c615c4cce9dbaf5
Signed-off-by: Dan Murphy <dmurphy@ti.com>
[Rebase on top of KK MR1]
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>